refactor(epicenter): move Id/Guid types and generators to shared/id.ts
Id and Guid are cross-cutting identity primitives used by dynamic/, filesystem/,
and static/ — not schema field definitions. Hoists customAlphabet calls to
module-level for minor perf improvement.
